backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests==2.31.0
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.0
radon==6.0.1
boto3==1.34.0
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.1
hiredis==2.3.2
neo4j==5.18.0
langchain==0.2.16
langchain-aws==0.1.17
python-jose[cryptography]==3.3.0
mangum==0.17.0
